Split Video rendering into component
